requirements.txt
fastapi
uvicorn
SQLAlchemy
aiosqlite
asyncpg
alembic
python-dotenv
python-jose
passlib
python-multipart
bcrypt